refactor(tpm2_protocol): decouple command building
Refactor the command serialization logic to eliminate the intermediate
buffer.

Responses don't possess this issue, as the auth area comes at the very end.
Thus, `TpmHeader` is extended to `TpmHeaderCommand` only for commands.

The new trait functions allow to refactor `tpm_build_command` to
orchestrate the build in a way that the auth area is nicely sandwiched
between handles and parameters.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>