refactor: integration test
Refactor `TpmDevice` to use a file descriptor and pass `CommandIO` instance
as a parameter to the subcommands.

With these changes applied, implement an integration test for the `import`
subcommand, which connects to a mocked TPM via socket. Implement a new
program called `mock-tpm` to provide fake TPM responses for subcommand
tests.

The test still blocks for yet to be discovered reason but since this does
not break anything else it is better to have a checkpoint.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>