fix(import): preserve parent object
`import` was incorrectly consuming its parent object from the pipeline.
This caused the parent to be missing from the final output.  Use
`io.get_active_object()` to get  reference to the parent object, and clone
it. This is the consistent with `seal` and some other commands.

This fix makes all tests pass again (including integration test).

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>