feat: migrate back to clap
It was good to work on more manual command-line for some time because now
there is a really good object structure, and command objects are easy to
perform integration testing. However, now it is time to bring back Clap to
make the maintaining less demanding.

This essentially only takes away `Command::parse()` and leaves other
parts intact and thus is not really step backwards.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>