feat(uri): add a new Uri type
URIs have become complicated enough that they need a proper parser and
test suite to validate its correctness. They are also the concept that
has evolved as a more intuitive substitute for the JSON pipeline. Since
they are at the center of this tool their validation ought to be first
class.

Write a new type using Pest, and a new suite, which validates the parse
with correct and incorrect strings.

After this commit, the plan is to refactor the application to use the
new type across the board for input data and only fully validated
`Uri` objects traverse through the system.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>